+++ Package.swift
@@ -12,7 +12,7 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/swiftlang/swift-docc-plugin", from: "1.4.3"),
         .package(url: "https://github.com/jpurnell/SwiftZIP.git", from: "0.6.0"),
-        .package(url: "https://github.com/jpurnell/SwiftExcelCore", exact: "0.4.0"),
+        .package(url: "https://github.com/jpurnell/SwiftExcelCore", exact: "0.5.0"),
     ],
     targets: [
         .target(